tests: fix -Wpointer-to-int-cast
../tests/test_virgl_resource.c:302:20: warning: cast from pointer to integer of different size [-Wpointer-to-int-cast]
  302 |   ck_assert_int_eq((unsigned long)priv, 0xabab);

Part-of: <https://gitlab.freedesktop.org/virgl/virglrenderer/-/merge_requests/1103>

+++ tests/test_virgl_resource.c
@@ -299,7 +299,7 @@ START_TEST(private_ptr)
   void *init_priv = (void*)0xabab;
   virgl_renderer_resource_set_priv(1, init_priv);
   void *priv = virgl_renderer_resource_get_priv(1);
-  ck_assert_int_eq((unsigned long)priv, 0xabab);
+  ck_assert_int_eq((uintptr_t)priv, 0xabab);
   testvirgl_fini_single_ctx();
 }
 END_TEST
